==> 2_add_user.ipynb <==
import pandas as pd
import re

# function to format movie title input, keeping only characters, numbers and spaces, and lowercase it
def clean(a):
    return re.sub(r'([^\s\w]|_)+', '', a).lower().title

movie_df = pd.read_csv("../data/raw/movies.csv")

# lowercase all titles in movie_df, getting ready for searching for user input
movie_df.title = movie_df.title.str.lower()
print(movie_df.head())

rating_df = pd.read_csv("../data/raw/movie_ratings_data_set.csv")
print(rating_df.head())

# create new user
# add 1 to max user_id to create new user_id
user_id = rating_df["user_id"].max() + 1
user_list = []
movie_list = []
rating_list = []

# movie_input = "Biker gangs@"
# rating_input = "3"

def rate_a_book():
    # get a movie title from user, and clean it
    print("What's the title of the movie you want to rate? (you can rate as many movies as you like)? For example, Biker Gangs.", "\n")
    movie_input = input()
    movie_input = clean(movie_input)

    # get a movie rating from user
    print(f"c", "\n")
    rating_input = int(input())

    # use movie_input to get movie_id
    movie_id = movie_df[movie_df["title"] == movie_input]["movie_id"].tolist()[0]

    user_list.append(user_id)
    movie_list.append(movie_id)
    rating_list.append(rating_input)

continue_or_not = "yes"
while continue_or_not == "yes":
    rate_a_book()
    print("Your rating is saved. Rate another book? yes/no")
    continue_or_not = input().lower()

new_rating_df = pd.DataFrame(
{
    "user_id": user_list,
    "movie_id": movie_list, 
    "value": rating_list
})
new_rating_df.to_csv("../data/not_raw/new_rating.csv", index=False)

all_rating_df = rating_df.append(new_rating_df).reset_index(drop=True)
all_rating_df.to_csv("../data/not_raw/all_rating.csv", index=False)

print("Great! We have all your ratings.")
print("Here are our recommendations for you:")
