==> 3_modeling.ipynb <==
import numpy as np
import pandas as pd
import pickle
import functions

# Load user ratings
raw_dataset_df = pd.read_csv('../data/not_raw/all_rating.csv')
# raw_dataset_df = raw_dataset_df.append({"user_id":101, "movie_id":2, "value":4}, ignore_index=True)
# print(raw_dataset_df.tail())

# Convert the running list of user ratings into a matrix
ratings_df = pd.pivot_table(raw_dataset_df, index='user_id', columns='movie_id', aggfunc=np.max)

# Normalize the ratings (center them around their mean)
normalized_ratings, means = functions.normalize_ratings(ratings_df.values)

# Apply matrix factorization to find the latent features
U, M = functions.low_rank_matrix_factorization(normalized_ratings,
                                                                    num_features=11,
                                                                    regularization_amount=1.1)

# Find all predicted ratings by multiplying U and M
predicted_ratings = np.matmul(U, M)

print(len(predicted_ratings))

# Add back in the mean ratings for each product to de-normalize the predicted results
predicted_ratings = predicted_ratings + means

# Save features and predicted ratings to files for later use
pickle.dump(U, open("user_features.dat", "wb"))
pickle.dump(M, open("product_features.dat", "wb"))
pickle.dump(predicted_ratings, open("predicted_ratings.dat", "wb" ))
pickle.dump(means, open("means.dat", "wb" ))
